requirements.txt
# My-Digirain Python 依赖

# 向量数据库
chromadb>=0.4.0

# 文档解析
pypdf>=3.0.0
epublib>=0.0.1
python-magic>=0.4.27

# HTTP 请求
requests>=2.28.0
httpx>=0.24.0

# 数据处理
numpy>=1.24.0
tqdm>=4.65.0

# 配置文件
pyyaml>=6.0

# 日志
loguru>=0.7.0
